# openapi/src/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: PayBox
  version: 1.0.0
  description: The Api collection for PayBox
  license:
    name: MIT license
    url: https://github.com/shawakash/payBox/blob/dev/LICENSE.md
servers:
  - url: http://localhost:8080
    description: Development server
...
paths:
...
  /client/:
    post:
      tags:
        - Client
      description: Client creation endpoint
      operationId: createClient
      requestBody:
        description: Client object that needs to be added
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClient"
      responses:
        "200":
          description: created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnCreatedClient"
        "409":
          description: Client already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Database Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error (jwt)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /txn/getMany:
    get:
      tags: 
        - Transaction
      description: Get Transactions based on networks and count endpoint
      operationId: getMany
      parameters:
        - in: header
          name: Authorization
          description: Access token for authentication
          required: true
          schema:
            type: string
        - in: query
          required: true
          name: txnQuery
          description: Network/s and counts Transaction query object
          schema:
            type: object
            properties:
              networks:
                oneOf:
                  - $ref: "#/components/schemas/Network"
                  - type: array
                    items:
                      $ref: "#/components/schemas/Network"
              count:
                type: integer
                description: Count representing the number of transactions required

      responses:
        "200":
          description: Some client transaction response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [ok]
                  txns:
                    $ref: "#/components/schemas/TxnArray"

        "403":
          description: Auth Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Database Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error (jwt)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /address/:
    post:
      tags:
        - Address
      description: Address creation endpoint
      operationId: createAddress
      parameters:
        - in: header
          name: Authorization
          description: Access token for authentication
          required: true
          schema:
            type: string
      requestBody:
        description: Address object that needs to be added
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressType"
      responses:
        "200":
          description: created address
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [ok]
                  id:
                    type: string
        
        "403":
          description: Address validation error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        
        "400":
          description: Incomplete request object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "409":
          description: Address already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Database Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error (jwt)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      tags: 
        - Address
      description: Address get endpoint
      operationId: getAddress
      parameters:
        - in: header
          name: Authorization
          description: Access token for authentication
          required: true
          schema:
            type: string
            
      responses:
        "302":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                      type: string
                      enum: [ok]
                  address:
                    $ref: "#/components/schemas/AddressType"
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                      type: string
                      enum: [ok]
                  address:
                    $ref: "#/components/schemas/AddressType"
        "404":
          description: Address Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Auth Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Database Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error (jwt)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
...
  schemas:
...
    ReturnCreatedClient:
      type: object
      properties:
        id:
          oneOf:
            - type: string
            - type: null
        jwt:
          type: string
        status:
          type: string
          enum:
            - ok
            - error
        username:
          type: string
        email:
          type: string
        mobile:
          oneOf:
            - type: number
            - type: null
        firstname:
          oneOf:
            - type: string
            - type: null
        lastname:
          oneOf:
            - type: string
            - type: null
        password:
          type: string
        address:
          oneOf: 
            - type: object
            - type: null
          properties:
            eth:
              oneOf:
                - type: null
                - type: string
            sol:
              oneOf:
                - type: null
                - type: string
            bitcoin:
              oneOf:
                - type: null
                - type: string
            usdc:
              oneOf:
                - type: null
                - type: string
    CreateClient:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        mobile:
          oneOf:
            - type: number
            - type: null
        firstname:
          oneOf:
            - type: string
            - type: null
        lastname:
          oneOf:
            - type: string
            - type: null
        password:
          type: string
      required:
        - username
        - email
        - password
...
    Error:
      type: object
      properties:
        status:
          type: string
          enum: [error]
        msg:
          type: string
...
    Network:
      type: string
      enum:
        - sol
        - eth
        - bitcoin
        - usdc
...
    TxnType:
      type: object
      properties:
        id:
          type: string
        time:
          type: string
        clientId:
          type: string
        blockTime:
          type: number
        amount:
          type: number
        fee:
          type: number
        from:
          type: string
        to:
          type: string
        postBalances:
          type: array
          nullable: true
          items:
            type: number
        preBalances:
          type: array
          nullable: true
          items:
            type: number
        recentBlockhash:
          type: string
        signature:
          type: array
          items:
            type: string
        network:
          $ref: "#/components/schemas/Network"
        slot:
          type: number
          nullable: true
        nonce:
          type: number
          nullable: true
        chainId:
          type: number
        cluster:
          oneOf:
            - $ref: "#/components/schemas/EthCluster"
            - $ref: "#/components/schemas/Cluster"
            - $ref: "#/components/schemas/BitcoinCluster"
            - $ref: "#/components/schemas/USDCCluster"
    EthCluster:
      type: string
      enum:
        - mainnet
        - ropsten
        - rinkeby
        - goerli
        - kovan
        - sepolia
    Cluster:
      type: string
      enum:
        - devnet
        - testnet
        - mainnet-beta
    BitcoinCluster:
      type: string
      enum:
        - mainnet
        - testnet
        - regtest
    USDCCluster:
      type: string
      enum:
        - ethereum_mainnet
        - ethereum_ropsten
        - ethereum_rinkeby
        - ethereum_testnet
        - ethereum_local
        - other
    TxnArray:
      type: array
      items:
        $ref: "#/components/schemas/TxnType"
    AddressType:
      type: object
      properties:
        eth:
          type: string
          pattern: "^0x[0-9a-fA-F]{40}$"  
        sol:
          type: string
          pattern: "^[0-9a-zA-Z]{32}$"   
        bitcoin:
          type: string
          nullable: true                  
        usdc:
          type: string
          nullable: true                  